MAIL-15 - remove last names from emails
This is primarily to reduce the data leak 'blast radius' if a donor or pledger accidentally makes
a typo and gives somebody else's valid email address. To keep the change low risk and make it easy
to revert if necessary, for now I am leaving the `*lastName` fields as required input for these
emails, since it is already sent and never blank. Mailer does not persist this data, which
is only transferred encrypted, so I think in this case it's OK to continue passing it to the app.