DON-1009: Share twig code between donation thanks mesage and regular giving confirmation message
This does change the way we show the first donation in the regular
giving message slightly. As far as I know the differences between that and
the ad-hoc donation message were not planned.

Probably most useful to review changes in the test snapshots.

'only' keyword stops twig auto-including all variables from the outer
template into the scope of the inner one, making the code more verbose
but more self-documenting.

I think this PR is enough to get the ticket ready for QA.